refactor: split ET.py into ET.py/ET2.py/_ET.py, add Rx/tx/etc free functions
ET2 (2D) previously shared a file with its 3D counterpart and their
common base class, which blocked exposing Rx/Ry/Rz/tx/ty/tz etc as bare
importable functions (tx/ty would collide between the two). Splits into:

- _ET.py: BaseET only (private/shared - can't import either concrete
  subclass without a cycle, since both depend on it).
- ET.py: the ET class + Rx/Ry/Rz/tx/ty/tz/SE3 as module-level names, so
  `from roboticstoolbox.ets.ET import *` pulls in exactly the 3D set.
- ET2.py: the ET2 class + R/tx/ty/SE2 as module-level names, same idea
  for 2D. tx/ty deliberately mean something different here than in ET.py
  - wildcard-importing both into one namespace isn't supported, by design.

isinstance(self, ET)/isinstance(self, ET2) inside BaseET's shared
__str__/__repr__ (which can't import either concrete subclass) are
replaced: __str__'s 3D-vs-2D compound-transform formatting now checks
`self.kind == "SE3"`/`"SE2"` directly (already guaranteed at that point -
no import needed, more robust than isinstance against subclassing).
__repr__'s class-name label now uses `self.__class__.__name__` (idiomatic
for a display label, and more correct for any future subclass than a
hardcoded "ET"/"ET2" ternary).

One landmine hit and fixed while writing this: `SE3 = ET.SE3` /
`SE2 = ET2.SE2` (the free-function aliases) would shadow the imported
spatialmath SE3/SE2 classes that the classmethod bodies reference by name
at call time - fixed by importing them as SE3T/SE2T instead.
