+++ src/roboticstoolbox/ets/__init__.py
@@ -1,4 +1,5 @@
-from roboticstoolbox.ets.ET import ET, ET2
+from roboticstoolbox.ets.ET import ET
+from roboticstoolbox.ets.ET2 import ET2
 from roboticstoolbox.ets.ETS import ETS, ETS2
 
 __all__ = [
