+++ src/roboticstoolbox/models/ETS/Planar2.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 
 import numpy as np
-from roboticstoolbox.ets.ET import ET2
+from roboticstoolbox.ets.ET2 import ET2
 from roboticstoolbox.ets.ETS import ETS2
 from roboticstoolbox.robot.Robot import Robot2
 from roboticstoolbox.robot.Link import Link2
