+++ src/roboticstoolbox/robot/Link.py
@@ -15,7 +15,9 @@ from typing import overload
 
 import roboticstoolbox as rtb
 from roboticstoolbox.ets.ETS import ETS, ETS2
-from roboticstoolbox.ets.ET import ET, ET2, BaseET
+from roboticstoolbox.ets.ET import ET
+from roboticstoolbox.ets.ET2 import ET2
+from roboticstoolbox.ets._ET import BaseET
 from warnings import warn
 
 from roboticstoolbox.tools.types import ArrayLike, NDArray
